requirements.txt
absl-py==2.4.0
action-msgs==2.0.3
action-tutorials-interfaces==0.33.9
action-tutorials-py==0.33.9
actionlib-msgs==5.3.6
aiofiles==25.1.0
ale-py==0.11.2
ament-cmake-test==2.5.5
ament-copyright==0.17.4
ament-cppcheck==0.17.4
ament-cpplint==0.17.4
ament-flake8==0.17.4
ament-index-python==1.8.2
ament-lint==0.17.4
ament-lint-cmake==0.17.4
ament-package==0.16.5
ament-pep257==0.17.4
ament-uncrustify==0.17.4
ament-xmllint==0.17.4
angles==1.16.1
asttokens==3.0.1
blinker==1.9.0
brax==0.14.1
builtin-interfaces==2.0.3
chex==0.1.91
click==8.3.1
cloudpickle==3.1.2
composition-interfaces==2.0.3
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.3.5
cv-bridge==4.1.0
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
demo-nodes-py==0.33.9
diagnostic-msgs==5.3.6
domain-coordinator==0.12.0
etils==1.13.0
example-interfaces==0.12.0
examples-rclpy-executors==0.19.7
examples-rclpy-minimal-action-client==0.19.7
examples-rclpy-minimal-action-server==0.19.7
examples-rclpy-minimal-client==0.19.7
examples-rclpy-minimal-publisher==0.19.7
examples-rclpy-minimal-service==0.19.7
examples-rclpy-minimal-subscriber==0.19.7
executing==2.2.1
Farama-Notifications==0.0.4
filelock==3.24.3
Flask==3.1.3
flask-cors==6.0.2
flax==0.12.4
fonttools==4.61.1
fsspec==2026.2.0
geometry-msgs==5.3.6
glfw==2.10.0
grpcio==1.78.1
gymnasium==1.2.3
humanize==4.15.0
image-geometry==4.1.0
importlib_resources==6.5.2
interactive-markers==2.5.5
ipython==9.10.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jax==0.9.1
jaxlib==0.9.1
jaxopt==0.8.5
jaxtyping==0.3.9
jedi==0.19.2
Jinja2==3.1.6
kiwisolver==1.4.9
laser-geometry==2.7.2
launch==3.4.10
launch-ros==0.26.11
launch-testing==3.4.10
launch-testing-ros==0.26.11
launch-xml==3.4.10
launch-yaml==3.4.10
lifecycle-msgs==2.0.3
logging-demo==0.33.9
lxml==6.0.2
map-msgs==2.4.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mediapy==1.2.6
message-filters==4.11.10
ml_collections==1.1.0
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
mujoco==3.5.0
mujoco-mjx==3.5.0
mypy_extensions==1.1.0
nav-msgs==5.3.6
networkx==3.6.1
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
opencv-python==4.13.0.92
opt_einsum==3.4.0
optax==0.2.6
orbax-checkpoint==0.11.33
orbax-export==0.0.8
osrf-pycommon==2.1.7
packaging==26.0
pandas==3.0.1
parso==0.8.6
pcl-msgs==1.0.0
pendulum-msgs==0.33.9
pexpect==4.9.0
pillow==12.1.1
-e git+https://github.com/google-deepmind/mujoco_playground.git@e8a41e7c9fee0f7fd3722922a0795d79854e0a7b#egg=playground
prompt_toolkit==3.0.52
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pygame==2.6.1
Pygments==2.19.2
PyOpenGL==3.1.10
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-qt-binding==2.2.2
PyYAML==6.0.3
qt-dotgraph==2.7.5
qt-gui==2.7.5
qt-gui-cpp==2.7.5
qt-gui-py-common==2.7.5
quality-of-service-demo-py==0.33.9
rcl-interfaces==2.0.3
rclpy==7.1.9
rcutils==6.7.5
resource-retriever==3.4.4
rich==14.3.3
rmw-dds-common==3.1.1
ros2action==0.32.8
ros2bag==0.26.9
ros2bag-mcap-cli==0.26.9
ros2bag-sqlite3-cli==0.26.9
ros2cli==0.32.8
ros2component==0.32.8
ros2doctor==0.32.8
ros2interface==0.32.8
ros2launch==0.26.11
ros2lifecycle==0.32.8
ros2multicast==0.32.8
ros2node==0.32.8
ros2param==0.32.8
ros2pkg==0.32.8
ros2plugin==5.4.4
ros2run==0.32.8
ros2service==0.32.8
ros2topic==0.32.8
rosbag2-interfaces==0.26.9
rosbag2-py==0.26.9
rosgraph-msgs==2.0.3
rosidl-adapter==4.6.7
rosidl-cli==4.6.7
rosidl-cmake==4.6.7
rosidl-generator-c==4.6.7
rosidl-generator-cpp==4.6.7
rosidl-generator-py==0.22.2
rosidl-generator-type-description==4.6.7
rosidl-parser==4.6.7
rosidl-pycommon==4.6.7
rosidl-runtime-py==0.13.1
rosidl-typesupport-c==3.2.2
rosidl-typesupport-cpp==3.2.2
rosidl-typesupport-fastrtps-c==3.6.3
rosidl-typesupport-fastrtps-cpp==3.6.3
rosidl-typesupport-introspection-c==4.6.7
rosidl-typesupport-introspection-cpp==4.6.7
rpyutils==0.4.2
rqt-action==2.2.1
rqt-bag==1.5.6
rqt-bag-plugins==1.5.6
rqt-console==2.2.2
rqt-graph==1.5.6
rqt-gui==1.6.3
rqt-gui-py==1.6.3
rqt-msg==1.5.2
rqt-plot==1.4.5
rqt-publisher==1.7.3
rqt-py-common==1.6.3
rqt-py-console==1.2.3
rqt-reconfigure==1.6.3
rqt-service-caller==1.2.2
rqt-shell==1.2.3
rqt-srv==1.2.3
rqt-topic==1.7.5
scipy==1.17.1
sensor-msgs==5.3.6
sensor-msgs-py==5.3.6
service-msgs==2.0.3
setuptools==69.5.1
shape-msgs==5.3.6
simplejson==3.20.2
six==1.17.0
sros2==0.13.5
stable_baselines3==2.7.1
stack-data==0.6.3
statistics-msgs==2.0.3
std-msgs==5.3.6
std-srvs==5.3.6
stereo-msgs==5.3.6
sympy==1.14.0
teleop-twist-keyboard==2.4.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
tensorstore==0.1.81
tf2-geometry-msgs==0.36.19
tf2-kdl==0.36.19
tf2-msgs==0.36.19
tf2-py==0.36.19
tf2-ros-py==0.36.19
tf2-sensor-msgs==0.36.19
tf2-tools==0.36.19
toolz==1.1.0
topic-monitor==0.33.9
torch==2.10.0
tqdm==4.67.3
traitlets==5.14.3
trajectory-msgs==5.3.6
treescope==0.1.10
trimesh==4.11.2
triton==3.6.0
turtlesim==1.8.3
type-description-interfaces==2.0.3
typing-inspect==0.9.0
typing_extensions==4.15.0
unique-identifier-msgs==2.5.0
uvloop==0.22.1
visualization-msgs==5.3.6
wadler_lindig==0.1.7
warp-lang==1.11.1
wcwidth==0.6.0
Werkzeug==3.1.6
zipp==3.23.0
